runner.go: Prompt caching
Currently, KV cache entries from a sequence are discarded at the
end of each processing run. In a typical chat conversation, this
results in each message taking longer and longer to process as
the entire history of the conversation needs to be replayed.

Prompt caching retains the KV entries as long as possible so
that we only need to process the newest message in the converation,
at least until there are too many simultaneous conversations and
something needs to be evicted.